Created an algorithm for two Matrices A and B
You have two matrices A and B of sizes n x n and need to compute their product C = A x B. 
Can you come up with an algorithm that can parallelize this computation to take advantage of multiple processors or cores? What is the theoretical limit of speedup you can achieve with this approach?